journald: forward message to syslog as it is (including null bytes)
systemd-journald currently strips whitespace from the end of a syslog
message body and leaves whitespace and the beginning of a message body.

In the past people have asked about possibility to preserve all
whitespace in log messages. Our position was that they should configure
forwarding to a traditional syslog implementation. However journald
strips trailing whitespace (part of message body) even before it
forwards message to a syslog daemon. Let's not do that and forward every
message as is.