basic: return number of attributes that were parsed
Previously cgroup_get_keyed_attribute() was all or
nothing. Either *all* wanted attributes were parsed and we returned
them or we returned none of them and reported an error (-ENXIO). Since
newer kernel versions add attributes that are generally not backported
to older versions it is more useful to parse what we can and let the
caller decide if he wants to make use of potentially incomplete data.