core/transaction: add taint flag informing user about dependency cycles
When dependency cycle is detected then all units on the cycle are marked
and we can provide information about this in systemctl status taint
string. This is useful in cases when cycle is detected when journal is
not running, e.g. start of the container when loggin only to console.